job_applications/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
boto3==1.40.8
botocore==1.40.8
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
cron-descriptor==1.4.5
cryptography==45.0.4
deprecation==2.1.0
Django==5.2.2
django-auditlog==3.1.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==3.2.0
drf-yasg==1.21.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
django-tenants==3.8.0
django-timezone-field==7.1
django-viewflow==2.2.11
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
fsspec==2025.5.1
gevent==25.5.1
geventhttpclient==2.3.4
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
locust==2.37.14
locust-cloud==1.26.3
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pipdeptree==2.27.0
platformdirs==4.3.8
postgrest==1.1.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pypdfium2==4.30.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-engineio==4.12.2
python-magic==0.4.27
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
realtime==2.5.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools<81
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
